# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: API para realizar consultas de buró de crédito con datos personales
    y financieros.
  title: CSMT-BIZ-BUREAU-OPERATION
  version: "0.1"
servers:
- description: Dev
  url: https://localhost
paths:
  /api/biz/v1/csmt/customer/bureau/score/post:
    post:
      description: Endpoint para realizar consultas de buró de crédito usando estructuras
        de datos complejas.
      parameters:
      - description: Identificador único de la traza para correlacionar eventos y
          logs a lo largo de la traza de una transacción.
        example: 123e4567-e89b-12d3-a456-426614174000
        explode: false
        in: header
        name: X-B3-TraceId
        required: false
        schema:
          type: string
        style: simple
      - description: Identificador único para el trabajo que realiza la llamada en
          el servicio actual.
        example: abc123def456
        explode: false
        in: header
        name: X-B3-SpanId
        required: false
        schema:
          type: string
        style: simple
      - description: "ID del Span del llamador, que es el Span ID del servicio o componente\
          \ que inició la llamada."
        example: 789abc456def
        explode: false
        in: header
        name: X-B3-ParentSpanId
        required: false
        schema:
          type: string
        style: simple
      - description: "Indica si la traza actual debe ser registrada o no. (1 = sí\
          , 0 = no)"
        example: "1"
        explode: false
        in: header
        name: X-B3-Sampled
        required: false
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsultaScoreRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultaScoreResponse'
          description: Éxito en la consulta del buró de crédito
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable content
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      summary: Realiza una consulta de score buró de crédito
      tags:
      - get-bureau-score
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: get-bureau-score
  /api/biz/v1/csmt/customer/bureau/detail/post:
    post:
      description: Endpoint para realizar consultas de buró de crédito usando estructuras
        de datos complejas.
      parameters:
      - description: Identificador único de la traza para correlacionar eventos y
          logs a lo largo de la traza de una transacción.
        example: 123e4567-e89b-12d3-a456-426614174000
        explode: false
        in: header
        name: X-B3-TraceId
        required: false
        schema:
          type: string
        style: simple
      - description: Identificador único para el trabajo que realiza la llamada en
          el servicio actual.
        example: abc123def456
        explode: false
        in: header
        name: X-B3-SpanId
        required: false
        schema:
          type: string
        style: simple
      - description: "ID del Span del llamador, que es el Span ID del servicio o componente\
          \ que inició la llamada."
        example: 789abc456def
        explode: false
        in: header
        name: X-B3-ParentSpanId
        required: false
        schema:
          type: string
        style: simple
      - description: "Indica si la traza actual debe ser registrada o no. (1 = sí\
          , 0 = no)"
        example: "1"
        explode: false
        in: header
        name: X-B3-Sampled
        required: false
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsultaDetalleRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultaDetalleResponse'
          description: Éxito en la consulta del buró de crédito
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable content
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      summary: Realiza una consulta de detalle buró de crédito
      tags:
      - get-bureau-detail
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: get-bureau-detail
  /api/biz/v1/csmt/customer/bureau/report/post:
    post:
      description: Endpoint para conciliaciones de área de riesgos
      parameters:
      - description: Identificador único de la traza para correlacionar eventos y
          logs a lo largo de la traza de una transacción.
        example: 123e4567-e89b-12d3-a456-426614174000
        explode: false
        in: header
        name: X-B3-TraceId
        required: false
        schema:
          type: string
        style: simple
      - description: Identificador único para el trabajo que realiza la llamada en
          el servicio actual.
        example: abc123def456
        explode: false
        in: header
        name: X-B3-SpanId
        required: false
        schema:
          type: string
        style: simple
      - description: "ID del Span del llamador, que es el Span ID del servicio o componente\
          \ que inició la llamada."
        example: 789abc456def
        explode: false
        in: header
        name: X-B3-ParentSpanId
        required: false
        schema:
          type: string
        style: simple
      - description: "Indica si la traza actual debe ser registrada o no. (1 = sí\
          , 0 = no)"
        example: "1"
        explode: false
        in: header
        name: X-B3-Sampled
        required: false
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsultaResumenRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultaResumenResponse'
          description: Éxito en la consulta del buró de crédito
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable content
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      summary: Realiza una consulta de las llamadas a buró de crédito en un periodo
        determinado
      tags:
      - get-bureau-report
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: get-bureau-report
components:
...
  schemas:
    ConsultaScoreRequest:
      example:
        consultaBuroRequest:
          Nombre:
            ApellidoMaterno: UNO
            SegundoNombre: ""
            EdadesDependientes: ""
            Sufijo: ""
            NumeroLicenciaConducir: ""
            NumeroRegistroElectoral: ""
            ApellidoAdicional: ""
            RFC: NOPJ800101
            Sexo: ""
            FechaNacimiento: "01011980"
            Prefijo: ""
            Residencia: ""
            ApellidoPaterno: PRUEBA
            NumeroCedulaProfesional: ""
            PrimerNombre: !!binary |-
              SlVBTg==
            Nacionalidad: MX
            NumeroDependientes: ""
            EstadoCivil: ""
            ClaveOtroPais: ""
            ClaveImpuestosOtroPais: ""
          Domicilios:
            Domicilio:
              Direccion2: ""
              DelegacionMunicipio: MIGUEL HIDALGO
              CodPais: MX
              Direccion1: JAIME BALMES 8
              NumeroTelefono: ""
              Ciudad: CD DE MEXICO
              CP: "11510"
              Estado: DF
              Extension: ""
              FechaResidencia: ""
              TipoDomicilio: ""
              ColoniaPoblacion: LOS MORALES POLANCO
              IndicadorEspecialDomicilio: ""
              Fax: ""
          Encabezado:
            TipoConsulta: I
            Idioma: ES
            TipoSalida: 2000-01-23T04:56:07.000+00:00
            TipoContrato: CC
            Version: "13"
            IdentificadorBuro: 2000-01-23
            ClavePais: MX
            NumeroReferenciaOperador: 122000854270 1 7086142637
            ClaveUnidadMonetaria: MX
            ClaveUsuario: XX99999999
            ProductoRequerido: "007"
            Password: 1ABCDefg
          Empleos:
            Empleo:
              Direccion2: ""
              DelegacionMunicipio: BENITO JUAREZ
              Cargo: ""
              CodPais: MX
              Direccion1: AVENIDA DIVISION DEL NORTE 1611
              NumeroTelefono: ""
              Ciudad: CIUDAD DE MEXICO
              NumeroEmpleado: ""
              CP: "03100"
              NombreEmpresa: DELEGACION BENITO JUAREZ
              Estado: DF
              Extension: ""
              BaseSalarial: ""
              ColoniaPoblacion: SANTA CRUZ ATOYAC
              FechaContratacion: ""
              Fax: ""
              FechaUltimoDiaEmpleo: ""
              ClaveMonedaSalario: ""
              Salario: ""
        configuracionBuro:
          forzarConsulta: true
          vigenciaMaxima: 60
      properties:
        configuracionBuro:
          $ref: '#/components/schemas/ConfiguracionBuroRequest'
        consultaBuroRequest:
          $ref: '#/components/schemas/ConsultaBuroRequest'
    ConsultaScoreResponse:
      example:
        fechaConsulta: 2012-04-23T18:25:43.511Z
        score: 600
        aprobado: true
      properties:
        aprobado:
          description: Estatus de la consulta al score
          example: true
          type: boolean
        score:
          description: Score obtendio del buró
          example: 600
          type: number
        fechaConsulta:
          description: Fecha de la consulta al score
          example: 2012-04-23T18:25:43.511Z
          format: date-time
          type: string
    ConsultaDetalleRequest:
      example:
        consultaBuroRequest:
          Nombre:
            ApellidoMaterno: UNO
            SegundoNombre: ""
            EdadesDependientes: ""
            Sufijo: ""
            NumeroLicenciaConducir: ""
            NumeroRegistroElectoral: ""
            ApellidoAdicional: ""
            RFC: NOPJ800101
            Sexo: ""
            FechaNacimiento: "01011980"
            Prefijo: ""
            Residencia: ""
            ApellidoPaterno: PRUEBA
            NumeroCedulaProfesional: ""
            PrimerNombre: !!binary |-
              SlVBTg==
            Nacionalidad: MX
            NumeroDependientes: ""
            EstadoCivil: ""
            ClaveOtroPais: ""
            ClaveImpuestosOtroPais: ""
          Domicilios:
            Domicilio:
              Direccion2: ""
              DelegacionMunicipio: MIGUEL HIDALGO
              CodPais: MX
              Direccion1: JAIME BALMES 8
              NumeroTelefono: ""
              Ciudad: CD DE MEXICO
              CP: "11510"
              Estado: DF
              Extension: ""
              FechaResidencia: ""
              TipoDomicilio: ""
              ColoniaPoblacion: LOS MORALES POLANCO
              IndicadorEspecialDomicilio: ""
              Fax: ""
          Encabezado:
            TipoConsulta: I
            Idioma: ES
            TipoSalida: 2000-01-23T04:56:07.000+00:00
            TipoContrato: CC
            Version: "13"
            IdentificadorBuro: 2000-01-23
            ClavePais: MX
            NumeroReferenciaOperador: 122000854270 1 7086142637
            ClaveUnidadMonetaria: MX
            ClaveUsuario: XX99999999
            ProductoRequerido: "007"
            Password: 1ABCDefg
          Empleos:
            Empleo:
              Direccion2: ""
              DelegacionMunicipio: BENITO JUAREZ
              Cargo: ""
              CodPais: MX
              Direccion1: AVENIDA DIVISION DEL NORTE 1611
              NumeroTelefono: ""
              Ciudad: CIUDAD DE MEXICO
              NumeroEmpleado: ""
              CP: "03100"
              NombreEmpresa: DELEGACION BENITO JUAREZ
              Estado: DF
              Extension: ""
              BaseSalarial: ""
              ColoniaPoblacion: SANTA CRUZ ATOYAC
              FechaContratacion: ""
              Fax: ""
              FechaUltimoDiaEmpleo: ""
              ClaveMonedaSalario: ""
              Salario: ""
        configuracionBuro:
          forzarConsulta: true
          vigenciaMaxima: 60
      properties:
        configuracionBuro:
          $ref: '#/components/schemas/ConfiguracionBuroRequest'
        consultaBuroRequest:
          $ref: '#/components/schemas/ConsultaBuroRequest'
    ConsultaDetalleResponse:
      example:
        consultaBuroResponse:
          Nombre:
            FechaNacimiento: "01011980"
            Prefijo: SR
            ApellidoMaterno: UNO
            Residencia: "1"
            ApellidoPaterno: PRUEBA
            PrimerNombre: !!binary |-
              SlVBTg==
            Nacionalidad: MX
            EstadoCivil: M
            RFC: NOPJ800101
            Sexo: M
          Cuentas:
            Cuenta:
              SaldoVencido: "0"
              SaldoActual: 10000+
              IdentificadorCAN: "02010100"
              NumeroCuentaActual: "989898121212"
              ModoReportar: M
              TipoCuenta: I
              ClaveUnidadMonetaria: MX
              CreditoMaximo: "50000"
              FechaUltimaCompra: "01012016"
              FormaPagoActual: 2000-01-23T04:56:07.000+00:00
              FechaUltimoPago: "01012016"
              FechaActualizacion: "20062016"
              FechaReporte: "09052016"
              NumeroPagos: "48"
              FechaAperturaCuenta: "01012016"
              HistoricoCAN: 11-2
              FechaMasAntiguaHistoricoCAN: "04092019"
              IndicadorTipoResponsabilidad: I
              FechaCancelacionCAN: "01012020"
              TipoContrato: AU
              FechaMasRecienteHistoricoCAN: "04122019"
              NombreOtorgante: BC 1TN
              LimiteCredito: "50000"
              FechaAperturaCAN: "01012019"
              IdentificadorCredito: "Y"
              ClaveOtorgante: XX99999999
              MontoPagar: 2000-01-23
              MontoUltimoPago: 2000-01-23
              FrecuenciaPagos: M
          ConsultasEfectuadas:
            ConsultaEfectuada:
              FechaConsulta: "05072016"
              IndicadorTipoResponsabilidad: I
              ClaveOtorgante: XX99999999
              TipoContrato: CC
              NombreOtorgante: BURO DE CREDITO
              ImporteContrato: "0"
          HawkAlertConsulta:
            HawkAlertC:
              Mensaje: DATOS GENERALES - COINCIDENCIA POR RFC
              CodigoClave: "020"
              FechaReporte: "31072014"
              TipoInstitucion: FINANCIERA
          Domicilios:
            Domicilio:
              DelegacionMunicipio: MIGUEL HIDALGO
              CodPais: MX
              Direccion1: INSUR SUR 1007
              FechaReporteDireccion: "04032016"
              Ciudad: CD DE MEXICO
              CP: "11230"
              Estado: DF
          HawkAlertBD:
            HawkAlertBD:
              Mensaje: DOM. PARTICULAR 3 - COINCIDENCIA POR DOMICILIO
              CodigoClave: "010"
              FechaReporte: "13082015"
              TipoInstitucion: FINANCIERA
          DeclaracionesCliente:
            DeclaracionConsumidor: DECLARATIVA GENERAL
          Encabezado:
            ClaveRetornoConsumidorSecundario: "0"
            NumeroControlConsulta: "1212121212"
            ClaveOtorgante: XX99999999
            IdentificadorBuro: 2000-01-23
            ClavePais: MX
            NumeroReferenciaOperador: 122000854270 1 7086142637
            ClaveRetornoConsumidorPrincipal: "1"
          ResumenReporte:
            ResumenReporte:
              TotalSaldosVencidosRevolventes: 2000-01-23
              NumeroMOPUR: 2000-01-23T04:56:07.000+00:00
              NumeroMOP96: 2000-01-23T04:56:07.000+00:00
              PctLimiteCreditoUtilizadoRevolventes: "2"
              CuentasClavesHistoriaNegativa: 2000-01-23
              NuevaDireccionReportadaUltimos60Dias: "N"
              MensajesAlerta: NNNYN
              TotalPagosPagosFijos: 2000-01-23
              CuentasPagosFijosHipotecas: 2000-01-23
              FechaAperturaCuentaMasAntigua: "01012010"
              CuentasRevolventesAbiertas: 2000-01-23
              TotalSolicitudesReporte: "99"
              FechaSolicitudReporteMasReciente: "05072016"
              ExistenciaDeclaracionesConsumidor: "N"
              NumeroMOP99: 2000-01-23T04:56:07.000+00:00
              NumeroMOP1: 2000-01-23T04:56:07.000+00:00
              NumeroMOP0: 2000-01-23T04:56:07.000+00:00
              TotalLimitesCreditoRevolventes: 2000-01-23T04:56:07.000+00:00
              NumeroMOP97: 2000-01-23T04:56:07.000+00:00
              NumeroMOP7: 2000-01-23T04:56:07.000+00:00
              NumeroMOP6: 2000-01-23T04:56:07.000+00:00
              CuentasCerradas: 2000-01-23
              NumeroTotalCuentasDespachoCobranza: 2000-01-23T04:56:07.000+00:00
              NumeroTotalSolicitudesDespachosCobranza: 2000-01-23T04:56:07.000+00:00
              FechaSolicitudMasRecienteDespachoCobranza: "00000000"
              NumeroMOP3: 2000-01-23T04:56:07.000+00:00
              NumeroMOP2: 2000-01-23T04:56:07.000+00:00
              NumeroMOP5: 2000-01-23T04:56:07.000+00:00
              NumeroMOP4: 2000-01-23T04:56:07.000+00:00
              TotalCreditosMaximosPagosFijos: 2000-01-23T04:56:07.000+00:00
              CuentasNegativasActuales: 2000-01-23
              TotalSaldosVencidosPagosFijos: "0"
              FechaIngresoBD: "23052013"
              FechaAperturaCuentaMasRecienteDespachoCobranza: "00000000"
              NumeroCuentas: 2000-01-23
              TotalSaldosActualesPagosFijos: 19469+
              CuentasDisputa: 2000-01-23T04:56:07.000+00:00
              TotalPagosRevolventes: 2000-01-23
              NumeroSolicitudesUltimos6Meses: "99"
              FechaAperturaCuentaMasReciente: "01012016"
              TipoMoneda: MX
              TotalSaldosActualesRevolventes: 4270+
          ScoreBuroCredito:
            ScoreBC:
              CodigoRazon: "13"
              nombreScore: BC SCORE
              ValorScore: 2000-01-23
              CodigoScore: "007"
          Empleos:
            Empleo:
              CodPais: MX
              Direccion1: ""
              FechaReportoEmpleo: "14032016"
              ModoVerificacion: M
              FechaVerificacion: "14032016"
              NombreEmpresa: JUBILADO
      properties:
        consultaBuroResponse:
          $ref: '#/components/schemas/ConsultaBuroResponse'
    ConsultaResumenRequest:
      example:
        FechaFin: yyyy/mm/dd
        FechaInicio: yyyy/mm/dd
      properties:
        FechaInicio:
          description: Fecha inicial de la consulta
          example: yyyy/mm/dd
          type: string
        FechaFin:
          description: Fecha inicial de la consulta
          example: yyyy/mm/dd
          type: string
    ConsultaResumenResponse:
      description: "Reporte de consultas a buro agrupado por ProductoRequerido, ClaveUsuario"
      items:
        $ref: '#/components/schemas/ConsultaResumenResponse_inner'
      type: array
    ConfiguracionBuroRequest:
      example:
        forzarConsulta: true
        vigenciaMaxima: 60
      properties:
        forzarConsulta:
          description: Forzar consulta a buró aunque exista una previa vigente
          example: true
          type: boolean
        vigenciaMaxima:
          description: Vigencia máxima en días
          example: 60
          type: number
      type: object
...
    ErrorResponse:
      properties:
        code:
          example: HTTP Code
          type: string
        message:
          example: 'Bad request / Unauthorized / Forbidden / Unprocessable content
            / Internal server error  '
          type: string
        detail:
          example: "Unexpected error occurred, please try again later."
          type: string
        timestamp:
          example: 2024-10-24T14:55:29Z
          format: date-time
          type: string
        type:
          example: InternalServerError
          type: string
      type: object
    ConsultaResumenResponse_inner_Reporte:
      description: "Reporte de consultas a buro agrupado por ProductoRequerido, ClaveUsuario"
      example:
        FechaFin: yyyy/mm/dd
        FechaInicio: yyyy/mm/dd
        Producto: Producto
        TotalConsultas: TotalConsultas
        ClaveUsuario: ClaveUsuario
      properties:
        FechaInicio:
          description: Fecha inicial de la consulta
          example: yyyy/mm/dd
          type: string
        FechaFin:
          description: Fecha inicial de la consulta
          example: yyyy/mm/dd
          type: string
        Producto:
          maxLength: 3
          type: string
        ClaveUsuario:
          maxLength: 3
          type: string
        TotalConsultas:
          maxLength: 3
          type: string
      type: object
...
    ConsultaResumenResponse_inner:
      example:
        Detalle:
        - Fecha: 2000-01-23
          Producto: Producto
          IdentificadorBuro: IdentificadorBuro
          ClaveUsuario: ClaveUsuario
          FolioBuro: FolioBuro
          ScoreBuroCredito:
          - CodigoRazon: "13"
            ValorScore: 2000-01-23
            CodigoScore: "007"
          - CodigoRazon: "13"
            ValorScore: 2000-01-23
            CodigoScore: "007"
        - Fecha: 2000-01-23
          Producto: Producto
          IdentificadorBuro: IdentificadorBuro
          ClaveUsuario: ClaveUsuario
          FolioBuro: FolioBuro
          ScoreBuroCredito:
          - CodigoRazon: "13"
            ValorScore: 2000-01-23
            CodigoScore: "007"
          - CodigoRazon: "13"
            ValorScore: 2000-01-23
            CodigoScore: "007"
        Reporte:
          FechaFin: yyyy/mm/dd
          FechaInicio: yyyy/mm/dd
          Producto: Producto
          TotalConsultas: TotalConsultas
          ClaveUsuario: ClaveUsuario
      properties:
        Reporte:
          $ref: '#/components/schemas/ConsultaResumenResponse_inner_Reporte'
        Detalle:
          description: Detalle de cada consulta que conforma el reporte
          items:
            $ref: '#/components/schemas/ConsultaResumenResponse_inner_Detalle_inner'
          type: array
      type: object
...